Fix typo. Refs #285
git-svn-id: https://svn.nexusformat.org/code/trunk@1644 ff5d1e40-2be0-497f-93bd-dc18237bd3c7

+++ include/napi.h
@@ -862,7 +862,7 @@ extern  NXstatus  NXsetcache(long newVal);
         NXstatus ( *nxopengroup) (NXhandle handle, CONSTCHAR *name, CONSTCHAR* NXclass);
         NXstatus ( *nxclosegroup)(NXhandle handle);
         NXstatus ( *nxmakedata64) (NXhandle handle, CONSTCHAR* label, int datatype, int rank, int64_t dim[]);
-        NXstatus ( *nxcompmakedata64) (NXhandle handle, CONSTCHAR* label, int datatype, int rank, int64_t dim[], int comp_typ, int bufsize[]);
+        NXstatus ( *nxcompmakedata64) (NXhandle handle, CONSTCHAR* label, int datatype, int rank, int64_t dim[], int comp_typ, int64_t bufsize[]);
         NXstatus ( *nxcompress) (NXhandle handle, int compr_type);
         NXstatus ( *nxopendata) (NXhandle handle, CONSTCHAR* label);
         NXstatus ( *nxclosedata)(NXhandle handle);
